move annotations to console
